changes to change callback to register.py and currently working on my local
